ChumpChangeServer/requirements.txt
aniso8601==8.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
plaid-python==7.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
